Move NotFound component to routes
