+++ src/router.tsx
@@ -2,7 +2,7 @@ import { createBrowserRouter } from "react-router-dom";
 import App from "./App";
 import Home from "./routes/Home";
 import Character from "./routes/Character";
-import NotFound from "./components/NotFound";
+import NotFound from "./routes/NotFound";
 
 const router = createBrowserRouter([
   {
